requirements.txt
faiss-cpu==1.10.0
hvac==2.3.0
numpy==2.2.3
sentence-transformers==3.4.1
torch==2.6.0
transformers==4.49.0
pyyaml==6.0.2
pydantic>=2.10.6
